# docs/openapi/chat-rest-api.yaml
openapi: 3.0.3
info:
  title: Chat Realtime REST API
  version: 1.0.0-mvp
  description: |
    REST API contract cho ung dung chat realtime co ban.

    Pham vi contract:
    - Auth bang JWT access token va refresh token
    - Ho so nguoi dung
    - Tim kiem user
    - Tao/mo hoi thoai 1-1
    - Danh sach hoi thoai
    - Lich su va gui tin nhan text
    - Danh dau da doc

    Realtime event contract duoc mo ta tai:
    docs/api-contract-chat-realtime-mvp.md
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/messages:
    get:
      tags:
        - Messages
      summary: Get message history for a conversation
      operationId: listMessages
      parameters:
        - $ref: '#/components/parameters/ConversationIdQueryParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/CursorParam'
      responses:
        '200':
          description: Message page sorted by createdAt ascending within a page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    post:
      tags:
        - Messages
      summary: Send a text message
      operationId: sendMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '201':
          description: Message created successfully. If this is the first message in a direct conversation, the target participant becomes visible in the conversation list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          description: Duplicate clientMessageId or conversation conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                code: MESSAGE_DUPLICATE
                message: Message with clientMessageId already exists
                timestamp: '2026-05-08T16:00:00Z'
                path: /api/v1/messages
  /api/v1/messages/read:
    post:
      tags:
        - Messages
      summary: Mark messages as read up to a message id
      operationId: markMessagesRead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarkReadRequest'
      responses:
        '200':
          description: Read position updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarkReadResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ConversationIdQueryParam:
      name: conversationId
      in: query
      required: true
      description: Persistent conversation id used to scope message operations.
      schema:
        type: integer
        format: int64
        minimum: 1
    LimitParam:
      name: limit
      in: query
      required: false
      description: Page size. Default 20, max 100.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    CursorParam:
      name: cursor
      in: query
      required: false
      description: Opaque cursor returned by previous page response.
      schema:
        type: string
...
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: VALIDATION_ERROR
            message: Request validation failed
            timestamp: '2026-05-08T16:00:00Z'
            path: /api/v1/auth/register
            fieldErrors:
              - field: email
                message: Email is invalid
    UnauthorizedError:
      description: Missing, invalid or expired access token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: UNAUTHORIZED
            message: Access token is invalid or expired
            timestamp: '2026-05-08T16:00:00Z'
            path: /api/v1/users/me
    ForbiddenError:
      description: Authenticated user is not allowed to access the resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: FORBIDDEN
            message: You are not allowed to access this resource
            timestamp: '2026-05-08T16:00:00Z'
            path: /api/v1/conversations/1001
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: NOT_FOUND
            message: Resource not found
            timestamp: '2026-05-08T16:00:00Z'
            path: /api/v1/conversations/1001
...
  schemas:
    ApiError:
      type: object
      required:
        - code
        - message
        - timestamp
        - path
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Request validation failed
        timestamp:
          type: string
          format: date-time
          example: '2026-05-08T16:00:00Z'
        path:
          type: string
          example: /api/v1/auth/register
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
          example: email
        message:
          type: string
          example: Email already exists
...
    PublicUser:
      type: object
      required:
        - id
        - username
        - displayName
      properties:
        id:
          type: integer
          format: int64
          example: 2
        username:
          type: string
          example: bob
        displayName:
          type: string
          example: Bob
        avatarUrl:
          type: string
          format: uri
          nullable: true
          example: https://cdn.example.com/avatars/bob.png
...
    Paging:
      type: object
      required:
        - limit
        - hasMore
      properties:
        limit:
          type: integer
          example: 20
        nextCursor:
          type: string
          nullable: true
          example: opaque-next-cursor
        hasMore:
          type: boolean
          example: true
...
    MessageStatus:
      type: string
      enum:
        - SENT
        - DELIVERED
        - READ
    MessageType:
      type: string
      enum:
        - TEXT
...
    MessageSender:
      $ref: '#/components/schemas/PublicUser'
    Message:
      type: object
      required:
        - id
        - conversationId
        - sender
        - content
        - messageType
        - status
        - createdAt
      properties:
        id:
          type: integer
          format: int64
          example: 9001
        clientMessageId:
          type: string
          nullable: true
          example: c8c4c246-0bb4-4d4f-94e7-512e8d6b6f6b
        conversationId:
          type: integer
          format: int64
          example: 1001
        sender:
          $ref: '#/components/schemas/MessageSender'
        content:
          type: string
          minLength: 1
          maxLength: 2000
          example: Hello Bob
        messageType:
          $ref: '#/components/schemas/MessageType'
        status:
          $ref: '#/components/schemas/MessageStatus'
        createdAt:
          type: string
          format: date-time
          example: '2026-05-08T16:02:00Z'
        deliveredAt:
          type: string
          format: date-time
          nullable: true
          example: '2026-05-08T16:02:03Z'
        readAt:
          type: string
          format: date-time
          nullable: true
          example: '2026-05-08T16:02:05Z'
    MessageListResponse:
      type: object
      required:
        - conversationId
        - items
        - paging
      properties:
        conversationId:
          type: integer
          format: int64
          example: 1001
        items:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        paging:
          $ref: '#/components/schemas/Paging'
    SendMessageRequest:
      type: object
      required:
        - conversationId
        - clientMessageId
        - content
      properties:
        conversationId:
          type: integer
          format: int64
          minimum: 1
          example: 1001
        clientMessageId:
          type: string
          minLength: 1
          maxLength: 100
          example: c8c4c246-0bb4-4d4f-94e7-512e8d6b6f6b
        content:
          type: string
          minLength: 1
          maxLength: 2000
          example: Hello Bob
        messageType:
          $ref: '#/components/schemas/MessageType'
      additionalProperties: false
    MarkReadRequest:
      type: object
      required:
        - conversationId
        - lastReadMessageId
      properties:
        conversationId:
          type: integer
          format: int64
          minimum: 1
          example: 1001
        lastReadMessageId:
          type: integer
          format: int64
          minimum: 1
          example: 9001
    MarkReadResponse:
      type: object
      required:
        - conversationId
        - lastReadMessageId
        - readAt
        - unreadCount
      properties:
        conversationId:
          type: integer
          format: int64
          example: 1001
        lastReadMessageId:
          type: integer
          format: int64
          example: 9001
        readAt:
          type: string
          format: date-time
          example: '2026-05-08T16:02:05Z'
        unreadCount:
          type: integer
          minimum: 0
          example: 0